requirements.txt
requests
pymongo[tls, srv]==4.4.*
gradio==3.45.*
python-dotenv